requirements.txt
Django==3.2.8
mysqlclient==2.1.1
djangorestframework==3.14.0